Removes ComicSiteModel
